feat:add-http-route, delete-http-route and get-http-route-list rpc commands implemented and bug fixed provided on both side (server & cli)

+++ cli/src/models/commands/echo_response.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct EchoResponse {
-    pub code: i64,
+    pub code: u16,
     pub message: String,
 }
